Move all download-tests to same module.
